preprocess continues
