+++ dla_cnn/desi/defs.py
@@ -1,0 +1,5 @@
+""" Definitions used throughout the package"""
+
+# Need to fill in TBD with the correct value
+REST_RANGE = [900, 1346, "TBD"]
+kernel = 400 # SDSS value -- UPDATE!!
