+++ dla_cnn/spectra_utils.py
@@ -1,0 +1,42 @@
+""" Methods for fussing with a spectrum """
+
+import numpy as np
+
+
+def get_lam_data(loglam, z_qso, REST_RANGE):
+    """
+    Generate wavelengths from the log10 wavelengths
+
+    Parameters
+    ----------
+    loglam: np.ndarray
+    z_qso: float
+    REST_RANGE: list
+        Lowest rest wavelength to search, highest rest wavelength,  number of pixels in the search
+
+    Returns
+    -------
+    lam: np.ndarray
+    lam_rest: np.ndarray
+    ix_dla_range: np.ndarray
+        Indices listing where to search for the DLA
+    """
+    lam = 10.0 ** loglam
+    lam_rest = lam / (1.0 + z_qso)
+    ix_dla_range = np.logical_and(lam_rest >= REST_RANGE[0], lam_rest <= REST_RANGE[1])
+
+    # ix_dla_range may be 1 pixels shorter or longer due to rounding error, we force it to a consistent size here
+    size_ix_dla_range = np.sum(ix_dla_range)
+    assert size_ix_dla_range >= REST_RANGE[2] - 2 and size_ix_dla_range <= REST_RANGE[2] + 2, \
+        "Size of DLA range assertion error, size_ix_dla_range: [%d]" % size_ix_dla_range
+    b = np.nonzero(ix_dla_range)[0][0]
+    if size_ix_dla_range < REST_RANGE[2]:
+        # Add a one to the left or right sides, making sure we don't exceed bounds on the left
+        ix_dla_range[max(b - 1, 0):max(b - 1, 0) + REST_RANGE[2]] = 1
+    if size_ix_dla_range > REST_RANGE[2]:
+        ix_dla_range[b + REST_RANGE[2]:] = 0  # Delete 1 or 2 zeros from right side
+    assert np.sum(ix_dla_range) == REST_RANGE[2], \
+        "Size of ix_dla_range: %d, %d, %d, %d, %d" % \
+        (np.sum(ix_dla_range), b, REST_RANGE[2], size_ix_dla_range, np.nonzero(np.flipud(ix_dla_range))[0][0])
+
+    return lam, lam_rest, ix_dla_range
